!store-imei: report IMEI in the log without requiring a subscriber
When store-imei is configured, a subscriber record is expected to exist
for receiving that IMEI storage - this mode is expected to be used
together with subscriber-create-on-demand.  However, what about
operation without store-imei?  Standard OsmoHLR still requires a
subscriber record to be present, even though the IMEI is only
emitted into log output.  Change this aspect, and also change the
log message format to make it friendlier to tools that will extract
IMSI-IMEI correspondence pairs from syslog.

Change-Id: If1fae788e943c6ffdb632751899ab5d3286b3428